+++ CHANGELOG.md
@@ -10,3 +10,4 @@
 - [feat:support stat reporting path aggregation.](https://github.com/Tencent/spring-cloud-tencent/pull/1608)
 - [feat:support instance detect.](https://github.com/Tencent/spring-cloud-tencent/pull/1617)
 - [fix: fix loadbalancer metadata type.](https://github.com/Tencent/spring-cloud-tencent/pull/1625)
+- [feat:support polaris server nearby backup LB.](https://github.com/Tencent/spring-cloud-tencent/pull/1627)